Move some functions in ceasiompyutils
