+++ ceasiompy/SU2MeshDef/su2meshdef.py
@@ -32,7 +32,7 @@ import shutil
 import numpy as np
 import pandas as pd
 
-import ceasiompy.utils.ceasiompyfunctions as ceaf
+from ceasiompy.utils.ceasiompyutils import get_results_directory, run_soft, aircraft_name
 import ceasiompy.utils.moduleinterfaces as mi
 
 from cpacspy.cpacsfunctions import (
@@ -546,10 +546,10 @@ def generate_mesh_def_config(tixi, wkdir, ted_uid, wing_uid, sym_dir, defl_list)
     """
 
     tigl = open_tigl(tixi)
-    aircraft_name = ceaf.aircraft_name(tixi)
+    ac_name = aircraft_name(tixi)
     DEFAULT_CONFIG_PATH = MODULE_DIR + "/files/DefaultConfig_v7.cfg"
     cfg = ConfigFile(DEFAULT_CONFIG_PATH)
-    config_dir_name = aircraft_name + "_TED_" + ted_uid
+    config_dir_name = ac_name + "_TED_" + ted_uid
     # TODO: add check or remove if already exist?
     os.mkdir(os.path.join(wkdir, "MESH", config_dir_name))
 
@@ -576,8 +576,8 @@ def generate_mesh_def_config(tixi, wkdir, ted_uid, wing_uid, sym_dir, defl_list)
 
     # TODO: is it the best way or should be pass as arg?
     mesh_dir = os.path.join(wkdir, "MESH")
-    su2_mesh_path = os.path.join(mesh_dir, aircraft_name + "_baseline.su2")
-    cfg["MESH_FILENAME"] = "../" + aircraft_name + "_baseline.su2"
+    su2_mesh_path = os.path.join(mesh_dir, ac_name + "_baseline.su2")
+    cfg["MESH_FILENAME"] = "../" + ac_name + "_baseline.su2"
 
     # Mesh Marker
     bc_wall_list, engine_bc_list = get_mesh_marker(su2_mesh_path)
@@ -617,7 +617,7 @@ def generate_mesh_def_config(tixi, wkdir, ted_uid, wing_uid, sym_dir, defl_list)
         cfg["DV_VALUE"] = str(defl / 1000)  # SU2 use 1/1000 degree...
 
         cfg["MESH_FILENAME"] = "_mesh_ffd_box.su2"
-        defl_mesh_name = aircraft_name + "_TED_" + ted_uid + "_defl" + str(defl) + ".su2"
+        defl_mesh_name = ac_name + "_TED_" + ted_uid + "_defl" + str(defl) + ".su2"
         if sym_dir:
             defl_mesh_name = "_" + defl_mesh_name
         cfg["MESH_OUT_FILENAME"] = defl_mesh_name
@@ -635,9 +635,7 @@ def generate_mesh_def_config(tixi, wkdir, ted_uid, wing_uid, sym_dir, defl_list)
             cfg["DV_VALUE"] = str(defl / 1000)  # SU2 use 1/1000 degree...
 
             cfg["MESH_FILENAME"] = defl_mesh_name
-            defl_mesh_sym_name = (
-                aircraft_name + "_TED_" + ted_uid + "_defl" + str(defl) + "_sym.su2"
-            )
+            defl_mesh_sym_name = ac_name + "_TED_" + ted_uid + "_defl" + str(defl) + "_sym.su2"
             cfg["MESH_OUT_FILENAME"] = defl_mesh_sym_name
 
             config_file_name = "ConfigROT_sym_defl" + str(defl) + ".cfg"
@@ -662,7 +660,8 @@ def generate_config_deformed_mesh(cpacs_path, cpacs_out_path, skip_config=False,
     """
 
     tixi = open_tixi(cpacs_path)
-    wkdir = ceaf.get_wkdir_or_create_new(tixi)
+
+    wkdir = get_results_directory("SU2Run")
 
     # Get SU2 mesh path
     su2_mesh_xpath = "/cpacs/toolspecific/CEASIOMpy/filesPath/su2Mesh"
@@ -674,8 +673,8 @@ def generate_config_deformed_mesh(cpacs_path, cpacs_out_path, skip_config=False,
         mesh_dir = os.path.join(wkdir, "MESH")
         if not os.path.isdir(mesh_dir):
             os.mkdir(mesh_dir)
-        aircraft_name = ceaf.aircraft_name(tixi)
-        su2_mesh_new_path = os.path.join(mesh_dir, aircraft_name + "_baseline.su2")
+        ac_name = aircraft_name(tixi)
+        su2_mesh_new_path = os.path.join(mesh_dir, ac_name + "_baseline.su2")
         shutil.copyfile(su2_mesh_path, su2_mesh_new_path)
         tixi.updateTextElement(su2_mesh_xpath, su2_mesh_new_path)
 
@@ -757,7 +756,7 @@ def run_mesh_deformation(tixi, wkdir):
         for cfg_file in sorted(cfg_file_list):
 
             if os.path.isfile(cfg_file):
-                ceaf.run_soft("SU2_DEF", cfg_file, ted_dir, nb_proc)
+                run_soft("SU2_DEF", cfg_file, ted_dir, nb_proc)
             else:
                 raise ValueError("Not correct configuration file to run!")
