+++ ceasiompy/utils/ceasiompyfunctions.py
@@ -24,13 +24,12 @@ import ceasiompy.__init__
 import os
 import shutil
 import datetime
-import platform
 import importlib
 from pathlib import Path
 from dataclasses import dataclass
-from contextlib import contextmanager
+from ceasiompy.utils.ceasiompyutils import change_working_dir
 
-from cpacspy.cpacsfunctions import get_value_or_default, open_tixi
+import ceasiompy.utils.moduleinterfaces as mi
 from ceasiompy.SettingsGUI.settingsgui import create_settings_gui
 from ceasiompy.Optimisation.optimisation import routine_launcher
 from ceasiompy.utils.configfiles import ConfigFile
@@ -44,7 +43,6 @@ log = get_logger(__file__.split(".")[0])
 MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
 LIB_DIR = Path(ceasiompy.__init__.__file__).parent
 
-SOFT_LIST = ["SU2_DEF", "SU2_CFD", "SU2_SOL", "mpirun.mpich", "mpirun"]
 
 OPTIM_METHOD = ["OPTIM", "DOE"]
 
@@ -77,15 +75,15 @@ class ModuleToRun:
 
         # Set default values
         self.is_settinggui = False
-        self.gui_related_module = []
+        self.gui_related_modules = []
         self.is_optim_module = False
         self.optim_method = None
         self.optim_related_modules = []
 
-    def define_settinggui(self, gui_related_module: list) -> None:
+    def define_settinggui(self, gui_related_modules: list) -> None:
 
         self.is_settinggui = True
-        self.gui_related_module = gui_related_module
+        self.gui_related_modules = gui_related_modules
 
     def define_optim_module(self, optim_method) -> None:
 
@@ -117,7 +115,7 @@ class ModuleToRun:
 
         if self.module_name == "SettingsGUI":
 
-            create_settings_gui(str(self.cpacs_in), str(self.cpacs_out), self.gui_related_module)
+            create_settings_gui(str(self.cpacs_in), str(self.cpacs_out), self.gui_related_modules)
 
         elif self.module_name in OPTIM_METHOD:
             # TODO: Implement optim module
@@ -245,10 +243,10 @@ class Workflow:
                 module_obj = ModuleToRun(module_name, self.current_wkflow_dir, cpacs_in)
 
             # Check if the module is SettingGUI
-            gui_related_modules = []
+            gui_related_moduless = []
             if module_name == "SettingsGUI":
-                gui_related_modules = self.get_gui_related_modules(self.module_to_run, m)
-                module_obj.define_settinggui(gui_related_modules)
+                gui_related_moduless = self.get_gui_related_moduless(self.module_to_run, m)
+                module_obj.define_settinggui(gui_related_moduless)
 
             skip_create_module = False
 
@@ -289,14 +287,15 @@ class Workflow:
 
             if module_obj.is_optim_module:
 
-                routine_launcher(module_obj.optim_method, module_obj.optim_related_modules)
+                with change_working_dir(self.current_wkflow_dir):
+                    routine_launcher(module_obj.optim_method, module_obj.optim_related_modules)
 
             else:
 
                 module_obj.run(self.current_wkflow_dir)
 
     @staticmethod
-    def get_gui_related_modules(module_list, idx) -> list:
+    def get_gui_related_moduless(module_list, idx) -> list:
         """Get modules list related to a specific SettingGUI module"""
 
         if "SettingsGUI" in module_list[idx + 1 :] and idx + 1 != len(module_list):
@@ -311,188 +310,10 @@ class Workflow:
 # =================================================================================================
 
 
-@contextmanager
-def change_working_dir(working_dir):
-    """Context manager to change the working directory just before the execution of a function."""
-
-    try:
-        cwd = os.getcwd()
-        os.chdir(working_dir)
-        yield
-    finally:
-        os.chdir(cwd)
-
-
-def get_results_directory(module_name: str) -> Path:
-    """Create (if not exists) and return the results directory for a module"""
-
-    if module_name not in get_submodule_list():
-        raise ValueError(f"Module '{module_name}' did not exit!")
-
-    specs = importlib.import_module(f"ceasiompy.{module_name}.__specs__")
-    results_dir = Path(Path.cwd(), specs.RESULTS_DIR)
-
-    if not results_dir.is_dir():
-        results_dir.mkdir(parents=True)
-
-    return results_dir
-
-
-def get_install_path(soft_check_list):
-    """Function to get installation paths a sorfware used in SU2
-
-    Function 'get_instal_path' check if the given list of software is installed,
-    it retruns a dictionnay of software with thier intallation paths.
-
-    Args:
-        soft_check_list (list): List of software to check installation path
-
-    Returns:
-        soft_dict (dict): Dictionary of software with their installation path
-
-    """
-
-    current_os = platform.system()
-
-    soft_dict = {}
-
-    for soft in soft_check_list:
-
-        # TODO: Check more and improve
-        if current_os == "Darwin":
-            log.info("Your OS is Mac")
-            # Run with MPICH not implemeted yet on mac
-            if "mpi" in soft:
-                install_path = ""
-            else:
-                install_path = "/Applications/SU2/" + soft
-
-        elif current_os == "Linux":
-            log.info("Your OS is Linux")
-            install_path = shutil.which(soft)
-
-        elif current_os == "Windows":
-            log.info("Your OS is Windows")
-            # TODO
-
-        else:
-            raise OSError("OS not recognized!")
-
-        if install_path:
-            log.info(soft + " is installed at: " + install_path)
-            soft_dict[soft] = install_path
-        elif "mpi" in soft:
-            log.warning(soft + " is not installed on your computer!")
-            log.warning("Calculations will be run on 1 proc only")
-            soft_dict[soft] = None
-        else:
-            raise RuntimeError(soft + " is not installed on your computer!")
-
-    return soft_dict
-
-
-# TODO make it more genearl, also for sumo and other
-def run_soft(soft, config_path, wkdir, nb_proc):
-    """Function run one of the existing SU2 software
-
-    Function 'run_soft' create the comment line to run correctly a SU2 software
-    (SU2_DEF, SU2_CFD, SU2_SOL) with MPI (if installed). The SOFT_DICT is
-    create from the SOFT_LIST define at the top of this script.
-
-    Args:
-        soft (str): Software to execute (SU2_DEF, SU2_CFD, SU2_SOL)
-        config_path (str): Path to the configuration file
-        wkdir (str): Path to the working directory
-
-    """
-
-    # Get installation path for the following softwares
-    SOFT_DICT = get_install_path(SOFT_LIST)
-
-    # mpi_install_path = SOFT_DICT['mpirun.mpich']
-    mpi_install_path = SOFT_DICT["mpirun"]
-
-    soft_install_path = SOFT_DICT[soft]
-
-    log.info("Number of proc available: " + str(os.cpu_count()))
-    log.info(str(nb_proc) + " will be used for this calculation.")
-
-    logfile_path = os.path.join(wkdir, "logfile" + soft + ".log")
-
-    # if mpi_install_path is not None:
-    #     command_line =  [mpi_install_path,'-np',str(nb_proc),
-    #                      soft_install_path,config_path,'>',logfile_path]
-
-    if mpi_install_path is not None:
-        command_line = [
-            mpi_install_path,
-            "-np",
-            str(int(nb_proc)),
-            soft_install_path,
-            config_path,
-            ">",
-            logfile_path,
-        ]
-    # elif soft == 'SU2_DEF' a disp.dat must be there to run with MPI
-    else:
-        command_line = [soft_install_path, config_path, ">", logfile_path]
-
-    log.info(f">>> Running {soft} on {nb_proc} proc")
-    log.info(f"    from {wkdir}")
-
-    with change_working_dir(wkdir):
-        os.system(" ".join(command_line))
-
-    log.info(f">>> {soft} End")
-
-    # TODO: try to use subprocess instead of os.system, how to deal with log file...?
-    # import subprocess
-    # p = subprocess.Popen(command_line, stdout=subprocess.PIPE)
-    # log_lines = p.communicate()[0]
-    # logfile = open(logfile_path, 'w')
-    # logfile.writelines(log_lines)
-    # logfile.close()
-
-
-def aircraft_name(tixi_or_cpacs):
-    """The function get the name of the aircraft from the cpacs file or add a
-        default one if non-existant.
-
-    Args:
-        cpacs_path (str): Path to the CPACS file
-
-    Returns:
-        name (str): Name of the aircraft.
-    """
-
-    # TODO: MODIFY this funtion, temporary it could accept a cpacs path or tixi handle
-    # check xpath
-    # *modify corresponding test
-
-    if isinstance(tixi_or_cpacs, str):
-
-        tixi = open_tixi(tixi_or_cpacs)
-
-        aircraft_name_xpath = "/cpacs/header/name"
-        name = get_value_or_default(tixi, aircraft_name_xpath, "Aircraft")
-
-        tixi.save(tixi_or_cpacs)
-
-    else:
-
-        aircraft_name_xpath = "/cpacs/header/name"
-        name = get_value_or_default(tixi_or_cpacs, aircraft_name_xpath, "Aircraft")
-
-    name = name.replace(" ", "_")
-    log.info("The name of the aircraft is : " + name)
-
-    return name
-
-
 # =================================================================================================
 #    MAIN
 # =================================================================================================
 
 if __name__ == "__main__":
 
-    log.info("Nothing to execute!")
+    print("Nothing to execute!")
