+++ ceasiompy/utils/tests/tests_ceasiompyfunctions/test_ceasiompyfunction.py
@@ -19,7 +19,8 @@ Python version: >=3.7
 import os
 import pytest
 from pathlib import Path
-from ceasiompy.utils.ceasiompyfunctions import ModuleToRun, Workflow, get_results_directory
+from ceasiompy.utils.ceasiompyfunctions import ModuleToRun, Workflow
+from ceasiompy.utils.ceasiompyutils import get_results_directory
 
 MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
 CPACS_PATH = os.path.join(MODULE_DIR, "D150_simple.xml")
@@ -43,7 +44,9 @@ class TestModuleToRun:
         assert self.module_works.cpacs_out.exists()
 
         assert not self.module_works.is_settinggui
-        assert not self.module_works.inculde_in_optim
+        assert self.module_works.gui_related_modules == []
+        assert not self.module_works.is_optim_module
+        assert self.module_works.optim_related_modules == []
         assert self.module_works.optim_method is None
 
     def test_module_name_error(self):
@@ -61,7 +64,7 @@ class TestModuleToRun:
         RELATED_MODULE = ["CPACS2SUMO", "SU2Run"]
         self.module_works.define_settinggui(RELATED_MODULE)
 
-        assert self.module_works.related_module == RELATED_MODULE
+        assert self.module_works.gui_related_modules == RELATED_MODULE
         assert self.module_works.is_settinggui
 
     def test_define_optim_module(self):
@@ -71,7 +74,7 @@ class TestModuleToRun:
         for optim_method in OPTIM_METHOD:
             self.module_works.define_optim_module(optim_method)
             assert self.module_works.optim_method == optim_method
-            assert self.module_works.inculde_in_optim
+            assert self.module_works.is_optim_module
 
         with pytest.raises(ValueError):
             self.module_works.define_optim_module("invalid_optim_method")
@@ -79,15 +82,15 @@ class TestModuleToRun:
     def test_create_module_wkflow_dir(self):
 
         # Remove dir from previous runs
-        if Path.joinpath(self.wkflow_test, "01_DOE_SU2Run").exists():
-            Path.joinpath(self.wkflow_test, "01_DOE_SU2Run").rmdir()
+        if Path.joinpath(self.wkflow_test, "01_DOE").exists():
+            Path.joinpath(self.wkflow_test, "01_DOE").rmdir()
 
         if Path.joinpath(self.wkflow_test, "02_SU2Run").exists():
             Path.joinpath(self.wkflow_test, "02_SU2Run").rmdir()
 
         self.module_works.create_module_wkflow_dir(1)
         assert self.module_works.module_wkflow_path.exists()
-        assert self.module_works.module_wkflow_path.stem == "01_DOE_SU2Run"
+        assert self.module_works.module_wkflow_path.stem == "01_DOE"
 
         self.module_works.optim_method = None
         self.module_works.create_module_wkflow_dir(2)
@@ -153,13 +156,16 @@ class TestWorkflow:
         assert self.workflow.current_wkflow_dir.exists()
         assert self.workflow.cpacs_path.exists()
 
-        assert len(list(self.workflow.current_wkflow_dir.iterdir())) == 8
+        assert len(list(self.workflow.current_wkflow_dir.iterdir())) == 7
 
-        for m, module in enumerate(self.MODULE_TO_RUN):
-            assert self.workflow.module_to_run_obj[m].module_name == module
+        assert self.workflow.module_to_run_obj[0].module_name == "SettingsGUI"
+        assert self.workflow.module_to_run_obj[1].module_name == "OPTIM"
+        assert self.workflow.module_to_run_obj[2].module_name == "PyTornado"
+        assert self.workflow.module_to_run_obj[3].module_name == "SettingsGUI"
+        assert self.workflow.module_to_run_obj[4].module_name == "PlotAeroCoefficients"
 
         assert self.workflow.module_to_run_obj[0].is_settinggui
-        assert self.workflow.module_to_run_obj[0].related_module == [
+        assert self.workflow.module_to_run_obj[0].gui_related_modules == [
             "SettingsGUI",
             "CPACS2SUMO",
             "CLCalculator",
@@ -169,16 +175,16 @@ class TestWorkflow:
             self.workflow.current_wkflow_dir, "01_SettingsGUI"
         )
 
-        assert self.workflow.module_to_run_obj[1].inculde_in_optim
+        assert self.workflow.module_to_run_obj[1].is_optim_module
         assert self.workflow.module_to_run_obj[1].optim_method == "OPTIM"
         assert self.workflow.module_to_run_obj[1].module_wkflow_path == Path.joinpath(
-            self.workflow.current_wkflow_dir, "02_OPTIM_CPACS2SUMO"
+            self.workflow.current_wkflow_dir, "02_OPTIM"
         )
 
-        assert self.workflow.module_to_run_obj[2].inculde_in_optim
-        assert self.workflow.module_to_run_obj[2].optim_method == "OPTIM"
+        assert not self.workflow.module_to_run_obj[2].is_optim_module
+        assert self.workflow.module_to_run_obj[2].optim_method is None
         assert self.workflow.module_to_run_obj[2].module_wkflow_path == Path.joinpath(
-            self.workflow.current_wkflow_dir, "03_OPTIM_CLCalculator"
+            self.workflow.current_wkflow_dir, "03_PyTornado"
         )
